third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-types.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id=popup popup>Popup</div>
<div id=hint popup=hint>Hint</div>
<div id=manual popup=manual>Async</div>
<div id=manual2 popup=manual>Async</div>

<script>
function assert_state_1(popupOpen,hintOpen,manualOpen,manual2Open) {
  assert_equals(popup.matches(':top-layer'),popupOpen,'popup open state is incorrect');
  assert_equals(hint.matches(':top-layer'),hintOpen,'hint open state is incorrect');
  assert_equals(manual.matches(':top-layer'),manualOpen,'manual open state is incorrect');
  assert_equals(manual2.matches(':top-layer'),manual2Open,'manual2 open state is incorrect');
}
test(() => {
  assert_state_1(false,false,false,false);
  popup.showPopUp();
  assert_state_1(true,false,false,false);
  hint.showPopUp();
  assert_state_1(true,true,false,false);
  manual.showPopUp();
  assert_state_1(true,true,true,false);
  manual2.showPopUp();
  assert_state_1(true,true,true,true);
  hint.hidePopUp();
  assert_state_1(true,false,true,true);
  popup.hidePopUp();
  assert_state_1(false,false,true,true);
  popup.showPopUp();
  hint.showPopUp();
  assert_state_1(true,true,true,true);
  popup.hidePopUp();
  assert_state_1(false,false,true,true);
  manual.hidePopUp();
  assert_state_1(false,false,false,true);
  manual2.hidePopUp();
  assert_state_1(false,false,false,false);
},'hints and manuals do not close popups');

test(() => {
  assert_state_1(false,false,false,false);
  hint.showPopUp();
  manual.showPopUp();
  manual2.showPopUp();
  assert_state_1(false,true,true,true);
  popup.showPopUp();
  assert_state_1(true,false,true,true);
  popup.hidePopUp();
  assert_state_1(false,false,true,true);
  manual.hidePopUp();
  manual2.hidePopUp();
  assert_state_1(false,false,false,false);
},'popups close hints but not manuals');
</script>

<div id=popup1 popup>Popup 1
  <div id=popup2 popup>Popup 2
    <p id=anchorid>Anchor</p>
    <div id=popup3 popup>Popup 3</div>
  </div>
</div>
<div id=hint2 popup=hint anchor=anchorid>Hint anchored to popup</div>
<script>
function assert_state_2(popup1Open,popup2Open,popup3Open,hintOpen) {
  assert_equals(popup1.matches(':top-layer'),popup1Open,'popup1 open state is incorrect');
  assert_equals(popup2.matches(':top-layer'),popup2Open,'popup2 open state is incorrect');
  assert_equals(popup3.matches(':top-layer'),popup3Open,'popup3 open state is incorrect');
  assert_equals(hint2.matches(':top-layer'),hintOpen,'hint2 open state is incorrect');
}
test(() => {
  assert_state_2(false,false,false,false);
  popup1.showPopUp();
  popup2.showPopUp();
  popup3.showPopUp();
  assert_state_2(true,true,true,false);
  hint2.showPopUp();
  assert_state_2(true,true,true,true);
  popup3.hidePopUp(); // Should close the hint
  assert_state_2(true,true,false,false);
  popup1.hidePopUp();
  assert_state_2(false,false,false,false);
},'hint, even with popup ancestor, closes with first popup');
</script>
